Refactoring to ease new device incorporation + start of pocket geiger implementation
