+++ src/app/components/photo/photo.component.ts
@@ -21,10 +21,8 @@ export class PhotoComponent implements ControlValueAccessor {
   @Input()
   title?: string;
 
-  // Function to call when the rating changes.
   private onChange = (option: any) => {};
 
-  // Function to call when the input is touched (when a star is clicked).
   private onTouched = () => {};
 
   constructor(private camera: Camera) {}
