+++ src/app/components/tag-list/tag-list.component.ts
@@ -21,10 +21,8 @@ export class TagListComponent implements ControlValueAccessor {
   @Input()
   title?: string;
 
-  // Function to call when the rating changes.
   private onChange = (option: any) => {};
 
-  // Function to call when the input is touched (when a star is clicked).
   private onTouched = () => {};
 
   registerOnChange(fn: (option: any) => void): void {
