+++ src/app/states/devices/abstract-device.service.ts
@@ -1,13 +1,13 @@
-import { BLE } from '@ionic-native/ble/ngx';
 import { Observable } from 'rxjs';
+import { Measure, Step } from '../measures/measure';
 import { AbstractDevice } from './abstract-device';
 
 export abstract class AbstractDeviceService<T extends AbstractDevice> {
-  constructor(protected ble: BLE) {}
-
   abstract getDeviceInfo(device: T): Observable<Partial<T>>;
 
   abstract saveDeviceParams(device: T): Observable<any>;
 
-  abstract startMeasureScan(device: T): Observable<any>;
+  abstract startMeasureScan(device: T, stopSignal: Observable<any>): Observable<Step>;
+
+  abstract computeRadiationValue(measure: Measure): number;
 }
