+++ src/app/states/devices/ble/device-atom-tag.ts
@@ -1,17 +1,12 @@
 import { _ } from '@biesbjerg/ngx-translate-extract/dist/utils/utils';
-import {
-  AbstractDevice,
-  DeviceParams,
-  DeviceParamsModel,
-  DeviceParamType,
-  DeviceType,
-  RawDevice
-} from './abstract-device';
+import { ApparatusSensorType, DeviceType } from '../abstract-device';
+import { DeviceParams, DeviceParamsModel, DeviceParamType } from '../device-params';
+import { AbstractBLEDevice, RawBLEDevice } from './abstract-ble-device';
 
-export class DeviceAtomTag extends AbstractDevice {
+export class DeviceAtomTag extends AbstractBLEDevice {
   readonly deviceType = DeviceType.AtomTag;
   apparatusVersion: string = DeviceType.AtomTag;
-  apparatusSensorType = 'Geiger';
+  apparatusSensorType = ApparatusSensorType.Geiger;
   apparatusTubeType = 'SBM-20';
 
   params: DeviceParams = {
@@ -29,7 +24,7 @@ export class DeviceAtomTag extends AbstractDevice {
     }
   };
 
-  constructor(rawDevice: RawDevice) {
+  constructor(rawDevice: RawBLEDevice) {
     super(rawDevice);
     if (rawDevice.advertising instanceof ArrayBuffer) {
       const data = new Uint8Array(rawDevice.advertising);
