+++ src/app/states/devices/ble/device-safe-cast.service.ts
@@ -1,23 +1,18 @@
 import { Injectable } from '@angular/core';
-import { BLE } from '@ionic-native/ble/ngx';
 import { Observable, of } from 'rxjs';
-import { Measure, Step } from '../measures/measure';
-import { DeviceSafeCast } from './device-safe-cast';
 import { bufferCount, filter, map, tap } from 'rxjs/operators';
+import { Measure, Step } from '../../measures/measure';
+import { AbstractBLEDeviceService } from './abstract-ble-device.service';
+import { DeviceSafeCast } from './device-safe-cast';
 
-// Todo add inheritance when angular issue fixed https://github.com/angular/angular/issues/24011
 @Injectable({
   providedIn: 'root'
 })
-export class DeviceSafeCastService /*extends AbstractDeviceService<DeviceSafeCast>*/ {
+export class DeviceSafeCastService extends AbstractBLEDeviceService<DeviceSafeCast> {
+  // Todo remove when angular issue fixed https://github.com/angular/angular/issues/24011
+  static ngInjectableDef = undefined;
   private service = 'ef080d8c-c3be-41ff-bd3f-05a5f4795d7f';
   private receiveCharacteristic = 'a1e8f5b1-696b-4e4c-87c6-69dfe0b0093b';
-
-  constructor(protected ble: BLE) {}
-
-  getDeviceInfo(): Observable<Partial<DeviceSafeCast>> {
-    return of({});
-  }
 
   computeRadiationValue(measure: Measure): number {
     if (measure.endTime) {
@@ -27,6 +22,14 @@ export class DeviceSafeCastService /*extends AbstractDeviceService<DeviceSafeCas
     } else {
       throw new Error('Incorrect measure : missing endTime');
     }
+  }
+
+  getDeviceInfo(device: DeviceSafeCast): Observable<Partial<DeviceSafeCast>> {
+    return of({});
+  }
+
+  saveDeviceParams(device: DeviceSafeCast): Observable<any> {
+    return of(null);
   }
 
   startMeasureScan(device: DeviceSafeCast, stopSignal: Observable<any>): Observable<Step> {
