--- src/app/states/devices/devices.service.spec.ts
@@ -1,15 +1,0 @@
-import { inject, TestBed } from '@angular/core/testing';
-
-import { DevicesService } from './devices.service';
-
-describe('DevicesService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [DevicesService]
-    });
-  });
-
-  it('should be created', inject([DevicesService], (service: DevicesService) => {
-    expect(service).toBeTruthy();
-  }));
-});
