+++ src/app/states/devices/usb/usb-devices.service.ts
@@ -1,0 +1,6 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class USBDevicesService {}
