--- src/app/states/measures/date.service.spec.ts
@@ -1,15 +1,0 @@
-import { inject, TestBed } from '@angular/core/testing';
-
-import { DateService } from './date.service';
-
-describe('DateService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [DateService]
-    });
-  });
-
-  it('should be created', inject([DateService], (service: DateService) => {
-    expect(service).toBeTruthy();
-  }));
-});
