+++ src/app/states/measures/measure.ts
@@ -1,9 +1,10 @@
 import { environment } from '../../../environments/environment';
+import { ApparatusSensorType } from '../devices/abstract-device';
 
 export class Measure {
   apparatusId?: string;
   apparatusVersion?: string;
-  apparatusSensorType?: string;
+  apparatusSensorType?: ApparatusSensorType;
   apparatusTubeType?: string;
   temperature?: number;
   value: number;
@@ -39,7 +40,7 @@ export class Measure {
   constructor(
     apparatusId: string | undefined,
     apparatusVersion: string | undefined,
-    apparatusSensorType: string | undefined,
+    apparatusSensorType: ApparatusSensorType | undefined,
     apparatusTubeType: string | undefined,
     deviceUuid: string,
     devicePlatform: string,
