--- src/app/states/measures/measures.service.spec.ts
@@ -1,15 +1,0 @@
-import { inject, TestBed } from '@angular/core/testing';
-
-import { PositionService } from './position.service';
-
-describe('PositionService', () => {
-  beforeEach(() => {
-    TestBed.configureTestingModule({
-      providers: [PositionService]
-    });
-  });
-
-  it('should be created', inject([PositionService], (service: PositionService) => {
-    expect(service).toBeTruthy();
-  }));
-});
